perf!: store source spans as u32 to shrink LabeledSpan
Change the internal storage and public API of `SourceSpan`/`SourceOffset`
(and therefore `LabeledSpan`) from `usize` to `u32`. This shrinks
`LabeledSpan` from 48 to 40 bytes.

Source files are bounded well below 4 GiB in practice, so a 32-bit byte
offset is sufficient. `ByteOffset` is now `u32`, getters return `u32`,
and the `(usize, usize)` / `Range<usize>` convenience conversions are
replaced with their `u32` equivalents.

BREAKING CHANGE: `SourceSpan`/`SourceOffset`/`LabeledSpan` offsets and
lengths are now `u32` instead of `usize`; offsets above `u32::MAX` are no
longer representable.

+++ tests/derive.rs
@@ -255,13 +255,13 @@ fn test_snippet_named_struct() {
         var1: SourceSpan,
         #[label = "var 2"]
         // Anything that's Clone + Into<SourceSpan> can be used here.
-        var2: (usize, usize),
+        var2: (u32, u32),
         #[label]
-        var3: (usize, usize),
+        var3: (u32, u32),
         #[label("var 4")]
-        var4: Option<(usize, usize)>,
+        var4: Option<(u32, u32)>,
         #[label]
-        var5: Option<(usize, usize)>,
+        var5: Option<(u32, u32)>,
     }
 }
 
@@ -299,9 +299,9 @@ fn test_snippet_enum() {
             #[label]
             var2: SourceSpan,
             #[label("var 3")]
-            var3: Option<(usize, usize)>,
+            var3: Option<(u32, u32)>,
             #[label]
-            var4: Option<(usize, usize)>,
+            var4: Option<(u32, u32)>,
         },
         #[diagnostic(code(foo::b))]
         B(
