+++ tests/narrated.rs
@@ -316,7 +316,7 @@ line4
 line5
 "#
     .to_string();
-    let len = src.len();
+    let len = src.len() as u32;
     let err = MyBad {
         src: NamedSource::new("bad_file.rs", src),
         highlight1: (0, len).into(),
@@ -379,7 +379,7 @@ line4
 line5
 "#
     .to_string();
-    let len = src.len();
+    let len = src.len() as u32;
     let err = MyBad {
         source: Inner(InnerInner),
         src: NamedSource::new("bad_file.rs", src),
